Rename shared/ -> core/; begin explicit .js extensions
