+++ packages/upscalerjs/src/browser/index.ts
@@ -1,20 +1,20 @@
-import { getUpscaler, } from '../shared';
+import { getUpscaler, } from '../core/index.js';
 import * as tf from '@tensorflow/tfjs';
-export * from '../shared';
-import { getUpscaleOptions, } from './args.browser';
-import { loadModel, } from './loadModel.browser';
+export * from '../core/index.js';
+import { getUpscaleOptions, } from './args.browser.js';
+import { loadModel, } from './loadModel.browser.js';
 import {
   getImageAsTensor,
   tensorAsBase64,
   checkValidEnvironment,
   Input,
-} from './image.browser';
+} from './image.browser.js';
 
 export default getUpscaler<typeof tf, Input>({
-    tf,
-    getUpscaleOptions,
-    loadModel,
-    getImageAsTensor,
-    tensorAsBase64,
-    checkValidEnvironment,
+  tf,
+  getUpscaleOptions,
+  loadModel,
+  getImageAsTensor,
+  tensorAsBase64,
+  checkValidEnvironment,
 });
