+++ packages/upscalerjs/src/core/constants.ts
