+++ packages/upscalerjs/src/core/errors-and-warnings.ts
