+++ packages/upscalerjs/src/core/image-utils.ts
