+++ packages/upscalerjs/src/core/index.ts
