+++ packages/upscalerjs/src/core/isLayersModel.ts
