+++ packages/upscalerjs/src/core/makeTick.test.ts
