+++ packages/upscalerjs/src/core/makeTick.ts
