+++ packages/upscalerjs/src/core/model-utils.ts
