+++ packages/upscalerjs/src/core/types.ts
