+++ packages/upscalerjs/src/core/upscale.ts
