+++ packages/upscalerjs/src/core/upscaler.ts
