+++ packages/upscalerjs/src/core/utils.ts
