+++ packages/upscalerjs/src/core/warmup.test.ts
