+++ packages/upscalerjs/src/core/warmup.ts
