+++ packages/upscalerjs/src/node-gpu/index.ts
@@ -1,14 +1,14 @@
-import { getUpscaler, } from '../shared';
+import { getUpscaler, } from '../core/index.js';
 import * as tf from '@tensorflow/tfjs-node-gpu';
-export * from '../shared';
-import { getUpscaleOptions, } from '../node/args.node';
-import { loadModel, } from '../node/loadModel.node';
+export * from '../core/index.js';
+import { getUpscaleOptions, } from '../node/args.node.js';
+import { loadModel, } from '../node/loadModel.node.js';
 import {
   getImageAsTensor,
   tensorAsBase64,
   checkValidEnvironment,
   Input,
-} from '../node/image.node';
+} from '../node/image.node.js';
 
 const Upscaler = getUpscaler<typeof tf, Input>({
   tf,
